# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MyStocks API
  version: 2.1.0
  description: |
    MyStocks 专业量化交易平台 API 规范。
    包含核心行情数据、技术分析、V3.1 风险管理系统及 Phase 2 信号监控模块。
  contact:
    name: 架构组
    email: arch@mystocks.com
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

  # --- 行情模块 ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    TokenResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: "bearer" }

    # --- 行情模型 ---
...